+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.51.0 (2021-06-11)
+
+#### :rocket: New Feature
+
+* In `derive` decorator added possibility to set default get/set accessors in trait classes.
+
 ## v3.50.0 (2021-06-07)
 
 #### :rocket: New Feature